conda_requirements.txt
# NB: these are requirements of code in this repo
# autoplot uses
affine
# pytest
beautifulsoup4
# don't download stuff
cartopy_offlinedata
# pytest
coverage
# one wxc script
ephem
# various places
fiona
# various places
gdal
# important
geopandas
# mrms script
imageio
# soft requirement for pandas
lxml
# pin down until bugs are fixed
matplotlib-base
# various places
metar
# duh
metpy>=1.0.0
# yup
netcdf4
# yup
numpy
# racoon
odfpy
# excel output
openpyxl
# yo
pandas
# mod-wsgi
paste
# PIL
pillow
# for downstream installs
pip
# for psql and friends
postgresql
# for pyiem
pyarrow
# lots of places
pygrib
# caching
pymemcache
# various places
pyproj
# various shapefile dumpers
pyshp>=2
# for testing
pytest
pytest-xdist
# various places
rasterio
# various places
rasterstats
# yo
requests
# important
scipy
# autoplots
seaborn
# needed for downstream pip install?
setuptools-markdown
# cow et al
shapely
# various dumpers
simplejson
sqlalchemy
# scripts
tqdm
# AHPS scripts
twisted
# need to use more
verde
# Excel
xlsxwriter
# Excel
xlwt
